Use AWS_USERNAME to set the role-session-name
Currently the role-session-name is being defined as a date. This means
that in Cloudtrail, API actions done via assume-role show up as a
timestamp.

However when you assume-role via the AWS Console, Amazon sets
role-session-name to be the username of the person who initially logged
in.

This changes assume-role to use the AWS_USERNAME field to closer match
how AWS does it.